Add checks for when guild or channel deleted

+++ src/structs/DeleteQueue.js
@@ -27,29 +27,31 @@ class DeleteQueue extends Serializable {
     for (let i=this.queue.length-1;i>=0;i--) {
       let message = this.queue[i];
       let success = true;
-      try {
-        await message.delete();
-      } catch(e) {
-        success = false;
-        // TODO: Add check for when bot will never be able to delete message
-        if (e.code === 10008 /* Unknown message */) {
-          debugLog(`Discord returning unknown message for [${message.id}] (${message.edits.length} edits) ${message.editedTimestamp-message.createdTimestamp}ms`);
-          let testMessage;
-          try {
-            testMessage = await message.channel.fetchMessage(message.id);
-          } catch(e) {
-            debugLog(`Couldn't fetch message [${message.id}]`);
-          } finally {
-            if (testMessage) this.queue[i] = testMessage;
-          }
+      if (!message.deleted) {
+        try {
+          await message.delete();
+        } catch(e) {
+          success = false;
+          // TODO: Add check for when bot will never be able to delete message
+          if (e.code === 10008 /* Unknown message */) {
+            debugLog(`Discord returning unknown message for [${message.id}] (${message.edits.length} edits) ${message.editedTimestamp-message.createdTimestamp}ms`);
+            let testMessage;
+            try {
+              testMessage = await message.channel.fetchMessage(message.id);
+            } catch(e) {
+              debugLog(`Couldn't fetch message [${message.id}]`);
+            } finally {
+              if (testMessage) this.queue[i] = testMessage;
+            }
 
-          /*
-          this.queue.pop(i);
-          delete this._queue[message.id];
-          deleted++;
-          */
-        } else {
-          debugLog(`Unable to delete old message [${message.id}] (${message.edits.length} edits) ${message.editedTimestamp-message.createdTimestamp}ms`, e);
+            /*
+            this.queue.pop(i);
+            delete this._queue[message.id];
+            deleted++;
+            */
+          } else {
+            debugLog(`Unable to delete old message [${message.id}] (${message.edits.length} edits) ${message.editedTimestamp-message.createdTimestamp}ms`, e);
+          }
         }
       }
       if (success) {
